Creatin the images data layer.
