+++ CatiLyfe.DataLayer/Models/Images/ImageAdapter.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CatiLyfe.DataLayer.Models.Images
+{
+    public enum ImageAdapter
+    {
+        AzureFile
+    }
+}
