+++ CatiLyfe.Database/Schemas/img/Index/pk_image.sql
@@ -1,0 +1,3 @@
+﻿ALTER TABLE img.images
+    ADD CONSTRAINT [pk_image]
+    PRIMARY KEY (slug)
