+++ CatiLyfe.Database/Schemas/img/Index/pk_links.sql
@@ -1,0 +1,3 @@
+﻿ALTER TABLE img.links
+    ADD CONSTRAINT [pk_links]
+    PRIMARY KEY (image, id)
