+++ CatiLyfe.Database/Schemas/img/Tables/images.sql
@@ -1,0 +1,6 @@
+﻿CREATE TABLE img.images
+(
+    id          INT           NOT NULL IDENTITY(1,1)
+   ,slug        NVARCHAR(128) NOT NULL
+   ,description NVARCHAR(256) NOT NULL
+)
