+++ CatiLyfe.Database/Schemas/img/img.sql
@@ -1,0 +1,1 @@
+﻿CREATE SCHEMA [img]
